create connector in index.js and delete web.js
