+++ package.json
@@ -7,7 +7,7 @@
     "build": "npm run clean && npm run build:server",
     "build:server": "./node_modules/.bin/babel server -d build --ignore __tests__",
     "start": "./node_modules/.bin/nf start",
-    "start:dev": "NODE_ENV=development ./node_modules/.bin/babel-watch -L server/web",
+    "start:dev": "NODE_ENV=development ./node_modules/.bin/babel-watch -L server",
     "test": "jest && npm run test:modules",
     "test:lint": "eslint server",
     "test:modules": "npm outdated --depth=0",
@@ -59,13 +59,12 @@
     "cors": "^2.8.1",
     "ejs": "^2.5.6",
     "express": "^4.15.2",
+    "foreman": "^2.0.0",
     "hogan.js": "^3.0.2",
     "hull": "^0.11.0",
     "jsforce": "^1.7.1",
-    "librato-node": "^4.1.0",
     "lodash": "^4.16.6",
     "newrelic": "^1.33.0",
-    "nf": "^0.0.3",
     "node-babel": "^0.1.2",
     "passport-forcedotcom": "^0.1.4",
     "rimraf": "^2.5.4",
