+++ server/index.js
@@ -1,29 +1,41 @@
-import librato from "librato-node";
+import express from "express";
 import Hull from "hull";
-import { name } from "../manifest.json";
+import server from "./server";
 
-if (process.env.LOGSTASH_HOST && process.env.LOGSTASH_PORT) {
-  const Logstash = require("winston-logstash").Logstash; // eslint-disable-line global-require
-  Hull.logger.add(Logstash, {
-    node_name: name,
-    port: process.env.LOGSTASH_PORT || 1515,
-    host: process.env.LOGSTASH_HOST
-  });
-  Hull.logger.info("start", { transport: "logstash" });
-} else {
-  Hull.logger.info("start", { transport: "console" });
+if (process.env.NEW_RELIC_LICENSE_KEY) {
+  Hull.logger.warn("starting newrelic with key: ", process.env.NEW_RELIC_LICENSE_KEY);
+  require("newrelic");
 }
 
-if (process.env.LIBRATO_TOKEN && process.env.LIBRATO_USER) {
-  librato.configure({
-    email: process.env.LIBRATO_USER,
-    token: process.env.LIBRATO_TOKEN
-  });
-  librato.on("error", (err) => {
-    console.error(err);
-  });
-  process.once("SIGINT", () => {
-    librato.stop(); // stop optionally takes a callback
-  });
-  librato.start();
+const config = {
+  hostSecret: process.env.SECRET || "BOOM",
+  port: process.env.PORT || 8082,
+  salesforce: {
+    oauth2: {
+      clientId: process.env.CLIENT_ID,
+      clientSecret: process.env.CLIENT_SECRET
+    }
+  },
+};
+
+const connector = new Hull.Connector(config);
+const app = express();
+
+connector.setupApp(app);
+
+function exitNow() {
+  Hull.logger.warn("Exiting now !");
+  process.exit();
 }
+
+function handleExit() {
+  Hull.logger.log("Exiting... waiting 30 seconds workers to flush");
+  setTimeout(exitNow, 30000);
+  server.exit().then(exitNow);
+}
+
+process.on("SIGINT", handleExit);
+process.on("SIGTERM", handleExit);
+
+server(app, config);
+connector.startApp(app);
